Move: pages폴더 안에있던 style 위치이동 및 로그인,회원가입 로그인지 접속제한
