+++ .babelrc
@@ -3,6 +3,11 @@
     "next/babel"
   ],
   "plugins": [
-    "babel-plugin-styled-components"
+    [
+      "styled-components",
+      {
+        "ssr": true
+      }
+    ]
   ]
 }
