+++ next.config.js
@@ -11,6 +11,7 @@ const nextConfig = {
       },
     ],
   },
+  pageExtensions: ['mdx', 'md', 'jsx', 'js', 'tsx', 'ts'],
 }
 
 module.exports = nextConfig
