+++ styles/errorStyle.ts
