+++ styles/mainStyle.ts
